requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
et-xmlfile==1.1.0
fastapi==0.95.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
httptools==0.5.0
idna==3.4
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
openpyxl==3.1.2
pandas==2.0.1
pinecone-client==2.2.1
protobuf==4.22.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
starlette==0.26.1
termcolor==2.3.0
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
yarl==1.9.2
